+++ bases/rsptx/assignment_server_api/assignment_builder/src/utils/exercise.ts
@@ -4,7 +4,7 @@ export const createExerciseId = (): string => {
     today.getFullYear() +
     (today.getMonth() + 1).toString().padStart(2, "0") +
     today.getDate().toString().padStart(2, "0");
-  const randInt = Math.floor(Math.random() * 1000);
+  const randInt = Math.floor(Math.random() * 10000);
 
   return `exercise_${date}_${randInt}`;
 };
